quest_parser: Fix case for constants
The generated onBNpcKill and onEObjHit functions were using the wrong
case for the constants they used. Ensure that they use the same case as
the declared constants, and reduce duplicate code.

Also, fix a typo in the onEObjHit function prototype which was causing
generated scripts to not compile properly, and revert the incorrect
addition (by me in an earlier commit) of "override".