Fix quest completion for questId > 2047 (67583)
It's currently impossible to permanently store quest completion flags
for quests beyond questId 2047 (67583) due to the code using a type
that's too small. This manifests in the correct quest completion packet
being sent, but as soon as you move to another zone the quest disappears
from the journal (and is probably replaced by another quest you didn't
ask for).

This fixes the issue by using an appropriately-sized index variable, and
also increases the flag storage so that all currently-available quests
can be stored.